+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Updated `standard_layout` function to contain class not id for each output block.
 * Added the `dim` slot to the list returned by the `plot_with_settings` module.
+* Added `style` argument to `verbatim_popup_srv` to control whether the content is styled using `styler::style_text`
 
 ### New features
 * Added a new module - `verbatim_popup`.